INS-1061: move storage helpers to separate file
