+++ ledger/storage/utils.go
@@ -1,0 +1,36 @@
+/*
+ *    Copyright 2018 Insolar
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package storage
+
+import (
+	"bytes"
+	"encoding/hex"
+)
+
+type bytes2hex []byte
+
+func (h bytes2hex) String() string {
+	return hex.EncodeToString(h)
+}
+
+func prefixkey(prefix byte, parts ...[]byte) []byte {
+	tail := bytes.Join(parts, nil)
+	k := make([]byte, len(tail)+1)
+	k[0] = prefix
+	_ = copy(k[1:], tail)
+	return k
+}
